Remove Gateway tests and move mappers into gateway
Integration test now covers the gateway tests.
As discussed in review with Zack.
